tests/wpt/web-platform-tests/dom/nodes/Document-contentType/contentType/contenttype_gif.html

<!DOCTYPE html>
<title>GIF document.contentType === 'image/gif'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe style="display:none"></iframe>
<script>
async_test(function() {
  var iframe = document.getElementsByTagName('iframe')[0];
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "image/gif");
  }), false);
  iframe.src = "../resources/t.gif";
});
</script>
